Client Done

+++ Client/src/com/silvercoffee/ecs/components/Position.java
@@ -4,31 +4,31 @@ package com.silvercoffee.ecs.components;
 import com.artemis.Component;
 
 public class Position extends Component{
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
     public Position() {
     }
 
-    public Position(int x, int y) {
+    public Position(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 }
